Checks performed all at once
Checks are now independently and automatically perfomed, as they set up during app initialisation. Checks service now holds major logic and AppService bootstraps the checks. ChecksService now exposes a method which uses a generic "_makeRequest" function. URLs were refactored to an array of objects with the data required to perform a specific check. It makes scaling easy - just add an object with endpoint to make request to, specify the HTTP verb (and payload if it is a POST request) and the "tokenRequired" flag if authentication is needed. N.B. :These changes were applied mostly due to the fact we need to serve data from DynamoDB on demand of client but client will always be receiving consistent data. Client side changes will be applied in further commits.