Payload separated from endpoints
As during root module init the app scans the table of endpoints it would be inefficient to add the same endpoint with different payload for the specific exchange and thus duplicate data. Therefore a separate table was created that contains endpoint URL, exchange specified and payload to make request with.