Complex check by creating and deleting order
The check is implemented by creating and immediately deleting after creation a trade order in order to notify if the service is down on some of the steps. This commit is temporary and the implementation will be refactored on further commits but for now the general idea and implementation for testing requests has been added.